tests refactor

+++ scrapper/src/main/java/edu/java/repositories/LinkRepositoryInterface.java
@@ -9,7 +9,7 @@ public interface LinkRepositoryInterface {
 
     List<Link> findAll();
 
-    Integer add(Link chat);
+    Long add(Link chat);
 
     Integer delete(Long id);
 }
